fix: expose  as top-level command (closes #241)
The README documents  as the way to scaffold a project, but
the CLI only registered it as . This change:

- Extracts the init action into a shared initAction function in
  packages/cli/src/commands/init.ts
- Registers  as a top-level command in index.ts
- Refactors ship.ts to import initAction from init.ts, keeping
   working as before

Both  and  now work identically.
